refactor(schedule): split ScheduleSettingsView.swift into per-struct files (#300)
ScheduleSettingsView.swift contained four `struct ... : View` types
in a single file (~300 lines total): the main settings view, the
schedule row, the create/edit form, and the wheel-style time picker
sheet. Move each non-main struct into its own file under a new
Views/Schedule/ directory so each view is one file, the main file
stays focused on the top-level surface.

- ScheduleSettingsView.swift: 300 → 98 lines (-202).
- New Views/Schedule/ScheduleRow.swift (62 lines).
- New Views/Schedule/ScheduleEditView.swift (122 lines).
- New Views/Schedule/TimePickerView.swift (44 lines).

Each new file gets a struct-level doc comment explaining the role
and the extraction context. Zero behavior change — each struct moved
verbatim, no logic touched.
